+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 - Print CLI tables using [comfy-table](https://crates.io/crates/comfy-table) ([#176](https://github.com/stackabletech/stackablectl/pull/176))
+- Bump kube to 0.77 ([#201](https://github.com/stackabletech/stackablectl/pull/201))
 
 ## [0.6.0] - 2022-10-28