+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 ### Changed
 
 - Print CLI tables using [comfy-table](https://crates.io/crates/comfy-table) ([#176](https://github.com/stackabletech/stackablectl/pull/176))
+- Bump kube to 0.77 ([#201](https://github.com/stackabletech/stackablectl/pull/201))
+- Bump go k8s client to 0.26.0 ([#202](https://github.com/stackabletech/stackablectl/pull/202))
+- Bump kube, k8s-openapi and tokio ([#205](https://github.com/stackabletech/stackablectl/pull/205))
 
 ## [0.6.0] - 2022-10-28